Refactored into junit test
